Add fuzzedTargets property to package model
It contains the targets that will need to be instrumented with
-sanitize=fuzzer. At the package description level, it is simply a list
of Strings. Later, it is a list that contains a subset of the actual
resolved targets (see: PackageGraphLoader.swift line 233).

The fuzzed targets can only be specified for version 4 packages.

I perform almost no error checking and wrote no tests.