Prototype custom build configurations + instrumentationSettings
Makes creation of custom build configurations possible, where a
BuildConfiguration is defined by its name alone and can refine a
previous configuration (e.g. “fuzz-debug” refines the config “debug”).

Adds an instrumentationSettings to the package manifest such that it is
possible to sanitize any resolved target depending on the build
configuration. These settings are local, that is, they will never be
applied if the package is compiled as a dependency.

- caching of manifests disabled
- Xcode project generation ignored
- tests failing
- build configuration probably not defined the way it should be
- instrumentationSettings contains minimum required functionality,
ideally it should be called localSettings and support all kinds of
settings and conditions
- BuildPlan probably does not handle the instrumentationSettings the
way it should